Docs: (1cbbd5b) feat(relay): check account balance is greater than a configured amount (#194)
* feat(relay): check account balance is greater than a configurable amount, rather than just >0

Right now, we relay messages for accouts with any balance. This fails to sybil-protect the relay
system as an attacker could send 1 wei to hundreds of accounts. This commit adds a field to the
relay server configuration specifying the minimum balance an account must have to get its messages
relayed.

* refactor(relay): edit minAccountBalance help message for consistency

* feat(relay): make minAccountBalance an uint64 instead of an int64

Co-authored-by: authcall <authcall@proton.me>